src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-tree-listener-clearance-expected.txt

CONSOLE ERROR: line 7: Uncaught TypeError: document.getElementById(...).createShadowRoot is not a function
CONSOLE ERROR: line 19: Uncaught TypeError: Cannot read property 'addEventListener' of undefined
FAIL Unexpected error: Uncaught TypeError: Cannot read property 'addEventListener' of undefined
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
